fix: Windows build - fix size_t to DWORD cast warning
Cast urlW.length() to DWORD to fix conversion warning.
Treating warnings as errors caused build failure.

+++ windows/runner/vpn_handler.cpp
@@ -241,7 +241,7 @@ bool VpnHandler::DownloadFile(const std::string& url, const std::string& destPat
   urlComponents.lpszUrlPath = urlPath;
   urlComponents.dwUrlPathLength = 1024;
   
-  if (!WinHttpCrackUrl(urlW.c_str(), urlW.length(), 0, &urlComponents)) {
+  if (!WinHttpCrackUrl(urlW.c_str(), static_cast<DWORD>(urlW.length()), 0, &urlComponents)) {
     WinHttpCloseHandle(hSession);
     return false;
   }
